feat(admin-dashboard): separate components + rbac
